Style changes as mentioned in review
