+++ src/Liminality/DiagramWriter.cs
@@ -31,10 +31,10 @@ namespace PSIBR.Liminality
         }
     }
 
-    public class MermaidDiagramWriter<TStateMachine> : DiagramWriter<TStateMachine>
+    public class MermaidStateDiagramWriter<TStateMachine> : DiagramWriter<TStateMachine>
         where TStateMachine : StateMachine<TStateMachine>
     {
-        public MermaidDiagramWriter(Graph<TStateMachine> graph) 
+        public MermaidStateDiagramWriter(Graph<TStateMachine> graph) 
             : base(graph)
         {
         }
@@ -53,32 +53,6 @@ namespace PSIBR.Liminality
             {
                 diagram.AddTransition(rootNode, node);
                 WriteNode(diagram, node);
-            }
-        }
-    }
-
-    public class MermaidStateDiagram : Diagram
-    {
-        private const string DiagramTypeToken = "stateDiagram-v2";
-        private const string Indent = "    ";
-        private const string GoesToToken = "-->";
-        private const string InitialStateToken = "[*]";
-
-        public MermaidStateDiagram()
-        {
-            AddSyntaxLine($"{DiagramTypeToken}");
-        }
-
-        public override void AddTransition(GraphNode? left, GraphNode? right)
-        {
-            string leftSyntax = left is null ? InitialStateToken : left.Name;
-            string rightSyntax = right is null ? InitialStateToken : right.Name;
-            string signalSyntax = left?.Condition is not null ? $":{left.Condition}" : string.Empty;
-
-            if (leftSyntax != rightSyntax && !string.IsNullOrWhiteSpace(signalSyntax))
-            {
-                var syntax = $"{Indent}{leftSyntax} {GoesToToken} {rightSyntax}{signalSyntax}";
-                AddSyntaxLine(syntax);
             }
         }
     }
